Section 119
Added a couple things. We added a vector input to the spawnprojectile function. Since we now have a location where the mouse is, we pass this location vector into the spawn projectile and we can get a rotation to that location so the projectile does not just fire straight anymore, it now fires towards the mouse location. We also added an input with the shift key while shooting a fireball, you dont have to be targeting an enemy, the cast firebolt spell will go off and the character will not additionally move, it will stay in place.