Enemy Health Bars
Created enemy health bars, started by adding a WidgetComponent to Enemy.h, and then utilizing OverlayWidgetController's delegate for OnAttributeChange, we broadcast changes to health and max health from BeginPlay. We set the widgetcontroller to Enemy, since it can be really any UObject, and then in Blueprints, subscribe to these delegates in the Graph. After making the UI bar of course, which is very simply just a progress bar and a color (red).