Damage Scaled by Curve Table
We changed the way damage is done, instead of raw editing the health value, damage is a seperate attribute that is subtracted from health. We can do calculations before the damage is done to the health, and can also scale it by an ability's level using a curve table. This is called a meta attribute (damage), and it is not replicated, only the final result is.