Finalizing the Hit React Ability
Added an array of common abilities to the ClassInfo data asset. We grant these abilities to the characters using the AuraBlueprintLibrary using GiveStartupAbilities function that we created. This function loops through all the abilities in the array that is in the data asset and addes them to the given abilities by their abilityspec. Once these abilities are granted, we activate them in the AttributeSet when damage is done, but first checking if the damage done was fatal or not.