Floating Damage Text
Setup a new UserWidget to handle floating text numbers, with an animation that makes it look like its drifting off into space. We then create a WidgetComponent class in C++ which handles setting the damage text. This class is spawned dynamically as an object either, on the server or as a remote call to the server if it is the client through the use of a reliable rpc function. This RPC function is called from attribute set, where the damage is passed through to it, and ultimately to the widget that sets the text number.