Aura Gameplay Effect Context
We created our own gameplay effect context, in which we put two new boolean variables that we can modify along the pipeline of applying effects to a target. In this case, we have Critical Hit boolean and a Blocked Damage boolean. So now we can use this information to modify the displayed text and other things besides just generically higher damage numbers.