Multiplayer Fixes
Fixed many multiplayer bugs that resulted from certain things being called at the wrong place or time in the code, resulting in either them not being replicated to the client properly (damage numbers) or not being destroyed at the right time (projectile). These were fixed checking various things, for instance at the beginning of the projectile overlap, making sure the context instigator was not the actor who shot the projectile, this fixed it overlapping with that instigator. Also the effect not being applied to the effectcauser, similiar scenerio. For the damage numbers, we had to make sure we were using the correct controller, and also that we werent displaying to the server if we were on the client, we only wanted the local controller. That fixed most of the more serious issues.