Behavior Tree Decorators
Fleshing out some of the conditionals for the behavior tree. If the enemy is melee, they get in close, if the enemy is ranged, they stay at some range. Also adjusted some of the Effect Actor settings, so that when enemies are running around we can chose to allow the effect to apply to them or not. This would be like, if they stand in fire, they can get burned... but can not pick up potions. To achieve this functionality we added a bool to the Effect actor class which configurable in Blueprint for each effect actor as to whether or not it applies to enemies.