Enemy AI using Environmental Query System
Using the EQS we made our ranged attackers modify their position based on if they can see Aura. We used a a custom made context in the EQS and made it so the enemy is constantly drawing a trace and getting the closest sphere that it can see aura at around a corner with the distance test score.