Melee Attack
Added Melee Attack ability, and granted this to specifically the warrior class enemy. Currently the warrior will get close, motion warp rotation to the character and then at the moment the spear is thrusted into the target, a debug sphere is displayed. To achieve the location of the spear tip, we needed to implement the native version of the getweapontiplocation function from the combat interface.