Multiple Melee Attack Sockets
In preperation for adding other melee characters of the warrior type, we needed to add some sockets to the hands of the Ghoul character. We achieved this by adding a struct to the Combat Interface which links together a Animation Montage with a gameplay tag. We then added an array of this struct to the character base class, which we can retrieve using a newly created function to do so (GetAttackMonages). This function is blueprintNative and is currently called in the Melee Attack gameplay ability, which we get a random asset of this array to use for the montage to be played. We then pass the associated gameplay tag into the Wait Event function so that the proper socket is being found, since the socket is 1:1 related to the montage currently playing. We then can pass that tag into the Get Socket location for a precise impact point during melee combat on whomever is attack whether it be the ghoul or the goblin spear.