Ranged Attacker with Sling Shot
Added a animationblueprint for the slingshot itself that makes the pouch follow the bone transform of the right hand socket of the owning goblin. This gives the pouch its slingshot elasticity and looks like a natural slingshot after playing the animation montage for it at the correct moment that the pouch should leave the hand of the goblin. We use a boolean variable that is modified on the montage event at the right moment to achieve this.