Melee Spear Impact Gameplay Cue
Made our first gameplay cue which allows replication of the sound and impact niagara system so that the client hears and sees the impact as well as the server. We need to do this because the ability is cast by an AI controller, and these are not replicated to begin with. Gameplay Cues are a system developed to deal with this limitation.