Received XP Information from PlayerState, to OverlayWidgetController
We bind to the delegate on the PlayerState that sends the XP information when it changes, and then use this information to calculate the appropriate bar percentage to display. We take the current XP total and by working with the Data Asset we can find what level we should land in with that much XP. Example, receive that XP is (5), Level 1 requires 0 xp, level 2 requires 10 xp, we'd be level 1.