Spell Menu UI
Made lots of additions to the spell menu, including a description for each spell. In addition to this  took a side quest to fix something in multiplayer with the enemy health bars. They were falling through the floor when the enemy died, added an interface function and called it in Multicast handle death in order to set the visibility to zero.