Getting Animatios setup For Electrocute Ability
Added a new class called AuraBeamSpell which gets information for the Electrocute ability. Also added a new state for when aura is currently in a shock loop which involved an interface function to retrieve the boolean off of Aura Character, it is set in the ability and the Event fires in Aura Character. It is a blueprintimplementable event in that way.